requirements.txt
requests==2.26.0
Flask==2.0.1
python-dateutil==2.8.2
modzy-sdk==0.5.6
deprecation==2.1.0